+++ middleman-core/lib/middleman-core/application.rb
@@ -1,3 +1,5 @@
+require 'socket'
+
 # Using Tilt for templating
 require 'tilt'
 
@@ -69,7 +71,7 @@ module Middleman
 
     # Which host preview should start on.
     # @return [Fixnum]
-    config.define_setting :host, '0.0.0.0', 'The preview server host'
+    config.define_setting :host, Socket.gethostname, 'The preview server host'
 
     # Which port preview should start on.
     # @return [Fixnum]