+++ CHANGELOG.md
@@ -3,10 +3,12 @@ master
 
 3.3.11
 ===
+* Add `srcset` option to `image_tag`. Also enables them in Markdown.
 * Add jruby to the list of envs requiring tzinfo-data.
 * Don't lookup resource for path if the path is absolute. Fixes #1195
 * Make preview host and port configurable in config.rb and also expose those variables to extensions which are curious. Closes #1477
 * Append assets hash to .woff2 files
+* Rack support has been broken since v3.3.9, fix that. #1501
 
 3.3.10
 ===